week2/requirements.txt
beautifulsoup4==4.14.3
chromadb==1.5.8
langchain==0.3.25
langchain-chroma==1.1.0
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==0.3.84
langchain-openai==0.3.25
langchain-text-splitters==0.3.11
lxml==6.0.4
openai==1.109.1
pypdf==6.10.2
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
streamlit==1.56.0
